+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - BREAKING: Refactored `RealtimeModel` and therefore, `OpenaiRealtimeRepository`.
 - BREAKING: Refactored `Message` class to be consistent with other APIs.
 - BREAKING: `SenderType` enum is now called `Role` to be consistent with other APIs.
+- Feat: `RealtimeModel` now has token usage stream.
 
 ## 3.10.3